refactor: separate item in MarketPlace component
