+++ src/pages/MarketPlace/index.tsx
@@ -1,0 +1,6 @@
+import MarketPlace from './MarketPlace'
+
+export * from './MarketPlace'
+export * from './MarketPlaceItem'
+
+export default MarketPlace
